fix admin-sdk-express running + use schema file
