+++ client/sandbox/admin-sdk-express/package.json
@@ -4,7 +4,8 @@
   "description": "Test script for admin SDK",
   "type": "module",
   "scripts": {
-    "dev": "node --watch --watch-preserve-output -r ts-node/register ./src/index.ts"
+    "dev": "node --watch --watch-preserve-output -r ts-node/register ./src/index.ts",
+    "dev:schema": "node --watch --watch-preserve-output --loader ts-node/esm ./src/with-schema.ts"
   },
   "dependencies": {
     "@instantdb/admin": "workspace:*",
